+++ src/wav.cpp
@@ -8,6 +8,16 @@
 #include "wav.h"
 
 /**
+ * How the public implementation (AudioFile.h) works:
+ * 
+ * 1. Define a 2D vector std::vector<std::vector<T>> that holds all of the samples. 
+ *    This vector is split into vector[channel][sampleIndex].
+ * 2. Use the header data correctly to do things
+ * 3. Validate the header data
+ * 4. Once done, read through the number of bytes specified in the header
+ *    and convert everything to a double
+
+/**
  * Read the input file and store WAV information to `header` and `buffer`.
  * THIS SHOULD BE THE CONSTRUCTOR SO THE WAV HEADER CAN'T BE READ BEFOREHAND
  * No error-checking or error output is currently provided.
@@ -20,7 +30,7 @@ void Wav::readFile(const std::string &fileName){
         this->fileName = fileName;
 
         wav_file.read((char*)&header, sizeof(wav_header)); //read into `header` by reading a number of bytes equivalent to the size of the struct
-        buffer = new signed short[header.buffer_size]; //allocate memory for the buffer
+        buffer = new signed int[header.buffer_size]; //allocate memory for the buffer
         wav_file.read((char*)buffer, header.buffer_size); //read remainder of file equal to buffer size into buffer
         wav_file.close();
     }
@@ -89,7 +99,7 @@ std::string Wav::getMetaData() const{
  * 
  * @return The buffer.
  */
-signed short* Wav::getBuffer(){
+signed int* Wav::getBuffer(){
     return buffer;
 }
